docs/module-6/solucao-exercicio/instrumentando-logs/book_store/pagamento/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.37
sqlalchemy-utils==0.41.2
psycopg2-binary==2.9.10
requests==2.32.3
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-instrumentation-sqlalchemy==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2